Clear the analyser findings in the VXI-11 code
CoreProcedure named values that fit in a byte while declaring a four-byte base;
the 32-bit wire field is produced in encodeCall, so the enum does not need it.
The test computed pointer offsets and byte counts by multiplying ints and
letting the result widen.

The local analysis harness had been running clang-tidy with relative paths,
which the project's HeaderFilterRegex ('.*/src/.*\.h$') never matches, so
findings in our own headers were invisible locally and only showed up in CI. It
passes absolute paths now and also covers tests/, which is what surfaced the two
in this commit.

+++ src/hal/include/peakemi/hal/Vxi11Protocol.h
@@ -23,8 +23,9 @@ inline constexpr std::uint32_t PortmapProtocolTcp = 6U;
 inline constexpr std::uint32_t CoreProgram = 0x0607AFU;
 inline constexpr std::uint32_t CoreVersion = 1U;
 
-/// Procedures of the VXI-11 core channel.
-enum class CoreProcedure : std::uint32_t
+/// Procedures of the VXI-11 core channel. The wire field is 32 bits wide, but
+/// that conversion happens in encodeCall; the enum only names the values.
+enum class CoreProcedure : std::uint8_t
 {
     CreateLink = 10,
     DeviceWrite = 11,
